pyproject.toml
[tool.poetry]
name = "pre-commit-github-actions"
version = "0.1.0"
description = "tutorial for using pre-commit together with github actions"
authors = ["pereza <alejandro.perez@socialbakers.com>"]
readme = "README.md"
packages = [{include = "pre_commit_github_actions"}]

[tool.poetry.dependencies]
python = "^3.9"
scrapy = "^2.9.0"
unidecode = "^1.3.6"
prefect = "^2.11.3"
prefect-gcp = "^0.4.5"
pandas-gbq = "^0.19.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.2.2"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
